Add viewer working with javafx and a little bit of code

+++ src/main/java/org/wingate/assj/tag/TagSetter.java
@@ -139,27 +139,27 @@ public class TagSetter {
 
                         // Inner tags
                         case "bord" -> {
-                            neo.setData(Float.parseFloat(m.group("param")));
+                            neo.setData(Double.parseDouble(m.group("param")));
                             tas.add(neo);
                         }
                         case "xbord" -> {
-                            neo.setData(Float.parseFloat(m.group("param")));
+                            neo.setData(Double.parseDouble(m.group("param")));
                             tas.add(neo);
                         }
                         case "ybord" -> {
-                            neo.setData(Float.parseFloat(m.group("param")));
+                            neo.setData(Double.parseDouble(m.group("param")));
                             tas.add(neo);
                         }
                         case "shad" -> {
-                            neo.setData(Float.parseFloat(m.group("param")));
+                            neo.setData(Double.parseDouble(m.group("param")));
                             tas.add(neo);
                         }
                         case "xshad" -> {
-                            neo.setData(Float.parseFloat(m.group("param")));
+                            neo.setData(Double.parseDouble(m.group("param")));
                             tas.add(neo);
                         }
                         case "yshad" -> {
-                            neo.setData(Float.parseFloat(m.group("param")));
+                            neo.setData(Double.parseDouble(m.group("param")));
                             tas.add(neo);
                         }
                         case "be" -> {
